Adjust double -> float to avoid the upcast

+++ src/Microsoft.DotNet.Wpf/src/PresentationCore/MS/internal/AnimatedTypeHelpers.cs
@@ -683,7 +683,7 @@ namespace MS.Internal.PresentationCore
 
         internal static bool IsValidAnimationValueSingle(float value)
         {
-            return double.IsFinite(value);
+            return float.IsFinite(value);
         }
 
         internal static bool IsValidAnimationValueSize(Size value)
